Filesystem: refactor-out of app + router context
Our data path / data dir is hardcoded, defined at compile time. There is
nothing flexible about it: once you start the router, you it will work
within the same data directory indefinitely (though there are some
run-time options that provide *some* flexibility where some of the
data within the directory can be found/redirected).

Initializing router context (as if the data dir were a mutable type)
is unnecessary and confusing. Ideally, we would *like* as much
flexibility as possible but, for now; this lends itself to breakage.

Also, this commit moves any filesystem related confusion out of app and
into core because of said reasons (and to remove redundancy). This is
not the *ideal* abstraction I had in mind for #330 but it is good
enough for now (and solves the issue at hand).

Run-time breakage was introduced in 95b35bf because of said reasons.

Closes #330
